server: move proxystore up one level
