+++ cmd/kpng/a2s.go
@@ -24,7 +24,7 @@ import (
 	"sigs.k8s.io/kpng/client/tlsflags"
 	"sigs.k8s.io/kpng/server/jobs/api2store"
 	"sigs.k8s.io/kpng/server/pkg/apiwatch"
-	"sigs.k8s.io/kpng/server/pkg/proxystore"
+	"sigs.k8s.io/kpng/server/proxystore"
 
 	"sigs.k8s.io/kpng/cmd/kpng/storecmds"
 )
