+++ cmd/kpng/storecmds/storecmds.go
@@ -28,7 +28,7 @@ import (
 	"sigs.k8s.io/kpng/server/jobs/store2api"
 	"sigs.k8s.io/kpng/server/jobs/store2file"
 	"sigs.k8s.io/kpng/server/jobs/store2localdiff"
-	"sigs.k8s.io/kpng/server/pkg/proxystore"
+	"sigs.k8s.io/kpng/server/proxystore"
 )
 
 type SetupFunc func() (ctx context.Context, store *proxystore.Store, err error)
