+++ server/jobs/api2store/api2store.go
@@ -28,7 +28,7 @@ import (
 
 	"sigs.k8s.io/kpng/api/localnetv1"
 	"sigs.k8s.io/kpng/server/pkg/apiwatch"
-	"sigs.k8s.io/kpng/server/pkg/proxystore"
+	"sigs.k8s.io/kpng/server/proxystore"
 )
 
 type Job struct {
