+++ server/jobs/kube2store/node-event-handler.go
@@ -20,7 +20,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 
 	localnetv1 "sigs.k8s.io/kpng/api/localnetv1"
-	proxystore "sigs.k8s.io/kpng/server/pkg/proxystore"
+	proxystore "sigs.k8s.io/kpng/server/proxystore"
 )
 
 const (
