+++ server/jobs/store2diff/store2diff.go
@@ -20,8 +20,8 @@ import (
 	"context"
 
 	"sigs.k8s.io/kpng/api/localnetv1"
-	"sigs.k8s.io/kpng/server/pkg/proxystore"
 	"sigs.k8s.io/kpng/server/pkg/server/watchstate"
+	"sigs.k8s.io/kpng/server/proxystore"
 )
 
 type Job struct {
