+++ server/jobs/store2globaldiff/store2globaldiff.go
@@ -23,8 +23,8 @@ import (
 	"sigs.k8s.io/kpng/api/localnetv1"
 	"sigs.k8s.io/kpng/client/lightdiffstore"
 	"sigs.k8s.io/kpng/server/jobs/store2diff"
-	"sigs.k8s.io/kpng/server/pkg/proxystore"
 	"sigs.k8s.io/kpng/server/pkg/server/watchstate"
+	"sigs.k8s.io/kpng/server/proxystore"
 )
 
 type Sink interface {
