+++ server/pkg/endpoints/node.go
@@ -20,7 +20,7 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	localnetv1 "sigs.k8s.io/kpng/api/localnetv1"
-	proxystore "sigs.k8s.io/kpng/server/pkg/proxystore"
+	proxystore "sigs.k8s.io/kpng/server/proxystore"
 )
 
 const hostnameLabel = "kubernetes.io/hostname"
