+++ server/pkg/endpoints/node_test.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	localnetv1 "sigs.k8s.io/kpng/api/localnetv1"
-	proxystore "sigs.k8s.io/kpng/server/pkg/proxystore"
+	proxystore "sigs.k8s.io/kpng/server/proxystore"
 )
 
 func ExampleForNodeWithTopology() {
