+++ server/pkg/proxy/server.go
@@ -32,7 +32,7 @@ import (
 	"k8s.io/klog/v2"
 
 	"sigs.k8s.io/kpng/client/tlsflags"
-	"sigs.k8s.io/kpng/server/pkg/proxystore"
+	"sigs.k8s.io/kpng/server/proxystore"
 )
 
 var (
